requirements.txt
numpy==1.21.6 
Pillow==9.5.0